improve all count schema feature to return long

+++ integration-test/src/test/java/org/apache/iotdb/db/it/schema/IoTDBCreateTimeseriesIT.java
@@ -158,7 +158,7 @@ public class IoTDBCreateTimeseriesIT {
         Statement statement = connection.createStatement();
         ResultSet resultSet = statement.executeQuery("count timeseries root.sg.**")) {
       while (resultSet.next()) {
-        int count = resultSet.getInt(1);
+        long count = resultSet.getLong(1);
         Assert.assertEquals(timeSeriesArray.length, count);
       }
     }
